+++ src/test/java/konkuk/thip/room/adapter/in/web/RoomPlayingDetailViewApiTest.java
@@ -1,8 +1,7 @@
 package konkuk.thip.room.adapter.in.web;
 
 import konkuk.thip.book.adapter.out.jpa.BookJpaEntity;
-import konkuk.thip.book.adapter.out.persistence.BookJpaRepository;
-import konkuk.thip.common.exception.InvalidStateException;
+import konkuk.thip.book.adapter.out.persistence.adapter.BookJpaRepository;
 import konkuk.thip.common.util.DateUtil;
 import konkuk.thip.common.util.TestEntityFactory;
 import konkuk.thip.room.adapter.out.jpa.CategoryJpaEntity;
