+++ src/test/java/konkuk/thip/room/adapter/in/web/RoomVerifyPasswordAPITest.java
@@ -2,7 +2,7 @@ package konkuk.thip.room.adapter.in.web;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import konkuk.thip.book.adapter.out.jpa.BookJpaEntity;
-import konkuk.thip.book.adapter.out.persistence.BookJpaRepository;
+import konkuk.thip.book.adapter.out.persistence.adapter.BookJpaRepository;
 import konkuk.thip.common.util.TestEntityFactory;
 import konkuk.thip.room.adapter.in.web.request.RoomVerifyPasswordRequest;
 import konkuk.thip.room.adapter.out.jpa.CategoryJpaEntity;
